Move category view urls to separate file
